reorganize project directories
