+++ README.md
@@ -29,3 +29,4 @@ This project will open at `http://localhost:5173`
 - 2.1 use fetch API to render card content V
 - 2.2 like button (locally store info)
 - 2.3 mock server ?
+- Fix card carousel functionality
