+++ vite-netflix-clone-project/src/scripts/api/fetchTest.ts
