+++ vite-netflix-clone-project/src/scripts/components/infiniteCarousel.ts
