+++ vite-netflix-clone-project/src/scripts/components/top10.ts
